third_party/WebKit/LayoutTests/fragmentation/become-unfragmented-with-float.html

<!DOCTYPE html>
<p>There should be green square below, and no red.</p>
<div id="multicol" style="position:relative; columns:2; column-fill:auto; height:60px; line-height:20px;">
    <div style="width:40px;">
        <div>
            <div style="height:30px;"></div>
            <div id="floater" style="float:left; width:100%; orphans:1; widows:1; background:red;">
                <div id="child1" style="background:green;"><br></div>
                <div id="child2" style="background:green;"><br></div>
            </div>
        </div>
    </div>
</div>
<div style="clear:both;"></div>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(() => {
    var multicol = document.getElementById("multicol");
    var floater = document.getElementById("floater");
    var child1 = document.getElementById("child1");
    var child2 = document.getElementById("child2");
    document.body.offsetTop;
    multicol.style.columns = "auto";
    multicol.style.height = "auto";
    assert_equals(floater.offsetTop, 30);
    assert_equals(child1.offsetTop, 30);
    assert_equals(child1.offsetHeight, 20);
    assert_equals(child2.offsetTop, 50);
    assert_equals(child2.offsetHeight, 20);
    assert_equals(floater.offsetHeight, 40);
}, "Become unfragmented with float at fragmentainer boundary");
</script>
